\name{BICZmiss}
\alias{BICZmiss}
\title{BIC of Z with missing values in X}
\usage{
  BICZmiss(X = X, Z = Z, M = NULL, Bic_vide_vect = NULL,
    BicOld = NULL, methode = 1, Zold = NULL, intercept = T,
    mixmod = NULL, nbit = 100, nbclustmax = 10,
    mixmodpenalty = T)
}
\description{
  BIC of Z with missing values in X
}

