\name{calcul_BIC_mixmod}
\alias{calcul_BIC_mixmod}
\title{calcul du BIC initial via package externe Mixmod (ou mclust)}
\usage{
  calcul_BIC_mixmod(X = X, nbclustmax = 10, bla = F,
    details = F, max = T, mclust = T)
}
\description{
  calcul du BIC initial via package externe Mixmod (ou
  mclust)
}

