% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_symbols.R
\name{clean_symbols}
\alias{clean_symbols}
\title{Remove and replace accents and special characters}
\usage{
clean_symbols(input)
}
\arguments{
\item{input}{The vector to clean}
}
\value{
a vector without special symbols and accents.
}
\description{
Remove and replace accents and special characters. 
Created to adapt names of dataframes coming from excel files with 
weird column names. Replacement is made by underscore symbol.
}
