% Generated by roxygen2 (4.0.1): do not edit by hand
\name{cleanYtest}
\alias{cleanYtest}
\title{Selection method based on p-values (coefficients)}
\usage{
cleanYtest(Y = Y, X = X, pvalmin = 0.05, bonferroni = F, A = NULL)
}
\arguments{
\item{Y}{the response variable}

\item{X}{the dataset of the covariates (without the response)}

\item{pvalmin}{the maximal bound for p-value associated to remaining coefficients}

\item{bonferroni}{boolean defining wether a Bonferroni correction is applied or not}

\item{A}{optional vector of coefficients to coerce some zeros}
}
\description{
Selection method based on p-values (coefficients)
}

