\name{cleanZlars}
\alias{cleanZlars}
\title{cleanZ by lars}
\usage{
  cleanZlars(Z = Z, X = X, methode = 1,
    mode = c("MSE", "BIC"), intercept = TRUE, K = NULL,
    groupe = NULL,
    select = c("lar", "lasso", "forward.stagewise", "stepwise", NULL))
}
\description{
  cleanZ by lars
}

