% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cleanZR2.R
\name{cleanZR2}
\alias{cleanZR2}
\title{To clean Z based on R2}
\usage{
cleanZR2(Z = Z, X = X, R2min = 0.4, methode = 1, adj = TRUE)
}
\arguments{
\item{Z}{binary adjacency matrix of the structure (size p)}

\item{X}{the dataset}

\item{R2min}{lower boundary for the structure (on R-squared value)}

\item{methode}{parameter for OLS (matrix inversion) methode_BIC  parameter for OLS (matrix inversion) 1:householderQr, 2:colPivHouseholderQr}

\item{adj}{boolean. Adjusted R-squared or classical one (if FALSE).}
}
\description{
To clean Z based on R2
}

