\name{cleanZtest}
\alias{cleanZtest}
\title{Clean Z's columns based on p-values (coefficients or global)}
\usage{
  cleanZtest(Z = Z, X = X, pvalmin = 0.05, methode = 1,
    adj = T, global = F, bonferroni = F)
}
\description{
  Clean Z's columns based on p-values (coefficients or
  global)
}

