\name{comparaison_struct}
\alias{comparaison_struct}
\title{To compare structures (Z)}
\usage{
comparaison_struct(trueZ = trueZ, Zalgo = Zalgo, all = TRUE,
  mode = "NULL")
}
\arguments{
  \item{trueZ}{first structure}

  \item{Zalgo}{second structure}

  \item{all}{boolean. Also compute the ratio for each
  stat.}

  \item{mode}{how to modify the structures before
  comparison. mode=c("NULL","hybrid","clique","sym")}
}
\description{
To compare structures (Z)
}

