\name{comparateur0_vect}
\alias{comparateur0_vect}
\title{compare zero in two coefficient vector (same size)}
\usage{
comparateur0_vect(vraiA = vraiA, Aalgo = Aalgo, taux = FALSE)

comparateur0_vect(vraiA = vraiA, Aalgo = Aalgo, taux = FALSE)
}
\arguments{
  \item{vraiA}{first vector}

  \item{Aalgo}{second vector}

  \item{taux}{boolean. Computes the ratio of each statistic
  or not.}
}
\description{
compare zero in two coefficient vector (same size)

compare 0 values in two vectors
}

