\name{comparateursign_vect}
\alias{comparateursign_vect}
\title{compare signs of the coefficients in two vectors}
\usage{
comparateursign_vect(vraiA = vraiA, Aalgo = Aalgo)
}
\arguments{
  \item{vraiA}{first vector}

  \item{Aalgo}{second vector}
}
\description{
compare signs of the coefficients in two vectors
}

