% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/compare_beta.R
\name{compare_beta}
\alias{compare_beta}
\title{compare signs of the coefficients in two vectors}
\usage{
compare_beta(truebeta = truebeta, hatbeta = hatbeta, taux = FALSE)
}
\arguments{
\item{truebeta}{first vector}

\item{hatbeta}{second vector}

\item{taux}{boolean. Computes the ratio of each selection statistic or not.}
}
\description{
compare signs of the coefficients in two vectors
}

