% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Conan.R
\name{Conan}
\alias{Conan}
\title{Conan the destructor}
\usage{
Conan(X = X, nbstep = Inf, std = FALSE, verbose = FALSE,
  coercing = NULL, Xout = TRUE)
}
\arguments{
\item{X}{the dataset (matrix)}

\item{nbstep}{number of cutting steps (may remove several rows or columns at each step)}

\item{verbose}{to print the result}

\item{coercing}{vector of the covariates to keep (names or index)}

\item{Xout}{boolean to export or not the reduced matrix (if not, indices are sufficient)}

\item{sd}{remove constant covariates}
}
\description{
Removes missing values (rows and column to obtain a large full matrix)
}

