\name{confint_coef}
\alias{confint_coef}
\title{confidence intervals on the coefficients estimated in a model or for proportions}
\usage{
confint_coef(modele = NULL, n = NULL, prop = NULL, mean = NULL,
  var = NULL, alpha = 0.05)
}
\description{
confidence intervals on the coefficients estimated in a
model or for proportions
}

