% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counting_events_group.R
\name{counting_events_group}
\alias{counting_events_group}
\title{Counts events}
\usage{
counting_events_group(
  events = events,
  group = NULL,
  base = NULL,
  prefix = NULL,
  suffix = NULL,
  clean = TRUE,
  keep = FALSE
)
}
\arguments{
\item{events}{The vector of events to count (qualitative values). 
One new variable is created for each unique value of events.}

\item{group}{qualitative vector that resets counting when its value changes.}

\item{base}{a dataframe to which counting can be added (cbind).}

\item{prefix}{(optional) prefix to the newly created variables names}

\item{suffix}{(optional) suffix to the newly created variables names}

\item{clean}{(boolean) to decide either or not variables names has to be 
cleaned (spaces, accents and special characters)}

\item{keep}{(boolean) to decide either or not events and group are to 
be added to the resulting data.frame}
}
\value{
a data.frame which number of rows is the length of events
}
\description{
creates a data.frame counting occurencies of specific events
in a given vector separated by named connex groups.
For example it allows to count occurencies of each category of a qualitative
 variable above in the dataset for each individual. 
 Grouping allows to reset the count each time another variable changes.
 
The idea is to create a cumulative history of what happened before 
current individual.

One new variable is created for each unique value of events
}
\examples{
events <- rep(c('a', 'b', 'c'), times = 20)
group <- rep(rep(c('A', 'B'), c(12, 8)), times = 3)
counting_events_group(events, group)

}
