% Generated by roxygen2 (4.0.1): do not edit by hand
\name{Estep}
\alias{Estep}
\title{Imputation of missing values knowing alpha (E step of the EM)}
\usage{
Estep(X = X, alpha = alpha, M = M, Z = Z, mixmod = mixmod, Zc = Zc)
}
\arguments{
\item{Ir}{the set of indices of the covariates on the left (computed once before EM to be efficient)}

\item{Zc}{profil (sum) colonne de Z}

\item{alpha}{matrix of the coefficients (p+1)xp}
}
\description{
Imputation of missing values knowing alpha (E step of the EM)
}

