\name{generateur_Y}
\alias{generateur_Y}
\title{generateur Y aleatoire}
\usage{
generateur_Y(X = X, Amax = NULL, sigma_Y = 10, positif = 0.6)
}
\arguments{
  \item{X}{the dataset}

  \item{Amax}{maximum number of non-zero coefficients}

  \item{sigma_Y}{the standard deviation of the noise}

  \item{positif}{the ratio of positive coefficients}
}
\description{
generateur Y aleatoire
}

