\name{generateurA_ou}
\alias{generateurA_ou}
\title{generateur de param?tres selon la structure}
\usage{
  generateurA_ou(Z = Z, tp1 = 1, tp2 = 1, tp3 = 1,
    positif = 0.6, lambdapois = 5, pb = 2, Amax = NULL,
    B = NULL)
}
\arguments{
  \item{pb}{kind of problem : 0=none, 1=simple, 2=strong}
}
\description{
  generateur de param?tres selon la structure
}

