% Generated by roxygen2 (4.0.1): do not edit by hand
\name{Gibbs_X_ij_IF}
\alias{Gibbs_X_ij_IF}
\title{Gibbs sampler for redundant covariates}
\usage{
Gibbs_X_ij_IF(Z = Z, X = X, p = p, mui = mui, sigmai = sigmai,
  alpha = alpha, mixmod = mixmod, j = j, components = components)
}
\arguments{
\item{X}{est la ligne concernée (X[i,])}

\item{components}{est le vecteur des classes pour i (vecteur creux de taille p avec pr elements non nuls)}

\item{mixmod}{is mixmod$details}
}
\description{
Gibbs sampler for redundant covariates
}

