\name{initial_cor3}
\alias{initial_cor3}
\title{calcul d'une structure initiale bas?e sur les corr?lations
correl=cor(X_appr)}
\usage{
  initial_cor3(correl = correl, p2max = Inf, rmax = Inf)
}
\description{
  calcul d'une structure initiale bas?e sur les
  corr?lations correl=cor(X_appr)
}

