% Generated by roxygen2 (4.0.1): do not edit by hand
\name{matplot_zone}
\alias{matplot_zone}
\title{draws matplot with conditionnal background for easier comparison of curves.}
\usage{
matplot_zone(x = x, y = y, col = 1:6, alpha = 0.2, what = which.min,
  ylim = NULL, xlim = NULL, type = "p", xlab = NULL, ylab = "NULL",
  main = NULL, lwd = lwd)
}
\description{
draws matplot with conditionnal background for easier comparison of curves.
}

