\name{meilleur_lars}
\alias{meilleur_lars}
\title{On estime les coefficients du meilleur mod?le obtenu (meilleur au sens du crit?re choisi)}
\usage{
  meilleur_lars(lars = lars, X = X, Y = Y,
    mode = c("MSE", "BIC"), intercept = TRUE, K = NULL,
    groupe = NULL, Amax = NULL)
}
\description{
  On estime les coefficients du meilleur mod?le obtenu
  (meilleur au sens du crit?re choisi)
}

