% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naive_model.R
\name{naive_model}
\alias{naive_model}
\title{How would it be if we were naive ?}
\usage{
naive_model(proba_1 = 0.8, effectif = 100, nb_1 = NULL)
}
\arguments{
\item{proba_1}{The ratio of 1 in the population (if nb_1 is NULL)}

\item{effectif}{The global effective of the population}

\item{nb_1}{The number of 1 in the population.If not NULL proba_1 is not read.default=NULL}
}
\description{
Describe the result of a naive binary discriminant model
}
