% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Numeric_Only.R
\name{Numeric_Only}
\alias{Numeric_Only}
\title{To clean non numeric values in a vector}
\usage{
Numeric_Only(x = NULL)
}
\arguments{
\item{x}{The vector to clean}
}
\description{
Replace all non numeric values in a vector by NA's and change the vector format to be numeric
}
