\name{ProbaZ}
\alias{ProbaZ}
\title{Probability of Z without knowing the dataset. It also gives the exact number of binary nilpotent matrices of size p.}
\usage{
ProbaZ(p = NULL, Z = NULL, proba = FALSE)
}
\arguments{
  \item{p}{the number of covariates}

  \item{Z}{the structure}
}
\description{
Probability of Z without knowing the dataset. It also gives
the exact number of binary nilpotent matrices of size p.
}

