\name{readY}
\alias{readY}
\title{a summary-like function}
\usage{
readY(A = A, labels = NULL, X = NULL, intercept = TRUE)
}
\arguments{
  \item{A}{coefficient vector}

  \item{labels}{name of the covariates}

  \item{X}{the dataset (named) if labels is null}

  \item{intercept}{boolean defining wether A contains an
  intercept or not}
}
\description{
a summary-like function
}

