\name{renet}
\alias{renet}
\title{correction for enet not working when X has only one variable}
\usage{
  renet(x, y, lambda = 0, max.steps, normalize = TRUE,
    intercept = TRUE, trace = FALSE,
    eps = .Machine$double.eps)
}
\description{
  correction for enet not working when X has only one
  variable
}

