% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_MSE.R
\name{report_MSE}
\alias{report_MSE}
\title{Quickly reports some MSE}
\usage{
report_MSE(real, prediction)
}
\arguments{
\item{real}{a numeric vector that contains the true values}

\item{prediction}{a numeric vector that contains the predicted values.}
}
\value{
a list containing the MSE values. 
RMSE is the root MSE (square root of the mean square error)
Relative is the mean of the relative errors (Root errors divided by the real values)
!standard_Deviation is the standard deviation of the root errors
}
\description{
Some MSE reporting to have an overview of the predictive quality of a model.
}
