% Generated by roxygen2 (4.0.1): do not edit by hand
\name{rforge}
\alias{rforge}
\title{Upgrades a package to the lastest version on R-forge}
\usage{
rforge(packages = c("CorReg"), repos = "http://R-Forge.R-project.org")
}
\arguments{
\item{packages}{vector of the name of the packages to upgrade}
}
\description{
Upgrades a package to the lastest version on R-forge
}

