\name{searchZmissR}
\alias{searchZmissR}
\title{MCMC}
\usage{
  searchZmissR(X = X, ini = NULL, maxit = 10^5, M = NULL,
    plot = F, BIC_vrai = NULL, Z_vrai = NULL,
    mixmod = mixmod, nbclustmax = 10, BIC_vide_vect = NULL,
    nett = T, mode = c("relax", "rejet"), MH = T,
    p2max = Inf, rmax = 5, nbit = 1)
}
\description{
  MCMC
}

