% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/showdata.R
\name{showdata}
\alias{showdata}
\title{show the missing values of a dataset}
\usage{
showdata(X = X, what = c("miss", "correl"))
}
\arguments{
\item{X}{the dataset to analyse}

\item{what}{indicates what to plot}
}
\description{
show the missing values of a dataset
}

