% Generated by roxygen2 (4.0.0): do not edit by hand
\name{WhoIs}
\alias{WhoIs}
\title{Give the partition implied by a structure}
\usage{
WhoIs(Z = Z, I3 = F, I2 = T, I1 = T)
}
\arguments{
\item{Z}{the structure (square matrix)}

\item{I1,I2,I3}{wanted outputs}
}
\description{
Give the partition implied by a structure
}

