% Generated by roxygen2 (4.0.0): do not edit by hand
\name{BicZ}
\alias{BicZ}
\title{Compute the BIC of a given structure}
\usage{
BicZ(X = X, Z = Z, Bic_null_vect = NULL, Bic_old = NULL, methode = 1,
  Zold = NULL, star = FALSE)
}
\arguments{
\item{X}{the dataset}

\item{Z}{binary adjacency matrix of the structure (size p)}

\item{Bic_null_vect}{the BIC of the null hypothesis (used for independent variables)}

\item{Bic_old}{BIC (vector) associated to Zold}

\item{Zold}{another structure with some common parts with Z (allows to compute only the differences)}

\item{methode}{parameter for OLS (matrix inversion) methode_BIC  parameter for OLS (matrix inversion) 1:householderQr, 2:colPivHouseholderQr}

\item{star}{boolean defining wether classical BIC or BIC* is computed}
}
\description{
Compute the BIC of a given structure
}

