\name{cleanZR2}
\alias{cleanZR2}
\title{To clean Z based on R2}
\usage{
cleanZR2(Z = Z, X = X, R2min = 0.4, methode = 1, adj = TRUE)
}
\arguments{
  \item{Z}{the structure}

  \item{X}{the dataset}

  \item{R2min}{lower boundary for the structure (on
  R-squared value)}

  \item{methode}{OLS parameter}

  \item{adj}{boolean. Adjusted R-squared or classical one
  (if FALSE).}
}
\description{
To clean Z based on R2
}

