% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanZtest.R
\name{cleanZtest}
\alias{cleanZtest}
\title{Clean Z's columns based on p-values (coefficients or global)}
\usage{
cleanZtest(Z = Z, X = X, pvalmin = 0.05, global = F,
  bonferroni = F)
}
\arguments{
\item{Z}{the binary matrix describing the sub-regression structure (as given by the "structureFinder" function).}

\item{X}{the dataset on which we have the sub-regression structure Z.}

\item{pvalmin}{the threshold on coefficients p-valuesto clean the structure.}

\item{global}{boolean. If TRUE the threshold is only on the F statistic for each sub-regression, not on each coefficients. So it will only remove entire sub-regressions.}

\item{bonferroni}{boolean to use bonferroni correction on the pvalmin parameter to avoid multiple testing issues.}
}
\description{
This function cleans the structure of correlations by setting to 0 the coefficients in the sub-regressions that are associated to a p-value below the "pvalmin" threshold.
}
