\name{comparaison_struct}
\alias{comparaison_struct}
\title{comparaison de structures}
\usage{
  comparaison_struct(vraiZ = vraiZ, Zalgo = Zalgo,
    tout = TRUE, mode = "NULL")
}
\arguments{
  \item{vraiZ}{first structure}

  \item{Zalgo}{second structure}

  \item{tout}{boolean. Also compute the ratio for each
  stat.}

  \item{mode}{how to modify the structures before
  comparison. mode=c("NULL","hybrid","clique","sym")}
}
\description{
  comparaison de structures
}

