\name{compare_beta}
\alias{compare_beta}
\title{compare signs of the coefficients in two vectors}
\usage{
compare_beta(truebeta = truebeta, hatbeta = hatbeta, taux = FALSE)
}
\arguments{
  \item{trueA}{first vector}

  \item{Aalgo}{second vector}

  \item{taux}{boolean. Computes the ratio of each selection
  statistic or not.}
}
\description{
compare signs of the coefficients in two vectors
}

