% Generated by roxygen2 (4.0.0): do not edit by hand
\name{compare_sign}
\alias{compare_sign}
\title{compare signs of the coefficients in two vectors}
\usage{
compare_sign(trueA = trueA, Aalgo = Aalgo)
}
\arguments{
\item{trueA}{first vector}

\item{Aalgo}{second vector}
}
\description{
compare signs of the coefficients in two vectors
}

