% Generated by roxygen2 (4.0.2): do not edit by hand
\name{density_estimation}
\alias{density_estimation}
\title{BIC of estimated  marginal gaussian mixture densities}
\usage{
density_estimation(X = X, nbclustmax = 10, nbclustmin = 1,
  verbose = FALSE, detailed = FALSE, max = TRUE, package = c("mclust",
  "Rmixmod", "rtkpp"), nbini = 20, matshape = FALSE, ...)
}
\arguments{
\item{X}{the dataset (matrix)}

\item{nbclustmax}{max number of clusters in the gaussian mixtures}

\item{nbclustmin}{min number of clusters in the gaussian mixtures}

\item{verbose}{verbose or not}

\item{detailed}{boolean to give the details of the mixtures found}

\item{max}{boolean. Use an heuristic to shrink nbclustmax according to the number of individuals in the dataset}

\item{package}{package to use (Rmixmod,mclust,rtkpp)}

\item{nbini}{number of initial points for Rmixmod}

\item{matshape}{boolean to give the detail in matricial shape}

\item{...}{additional parameters}
}
\description{
Estimates the density of each covariates with gaussian mixture models and then gives the associated BIC.
}

