% Generated by roxygen2 (4.0.1): do not edit by hand
\name{density_estimation}
\alias{density_estimation}
\title{BIC of estimated  marginal gaussian mixture densities}
\usage{
density_estimation(X = X, nbclustmax = 10, nbclustmin = 1,
  verbose = FALSE, detailed = FALSE, max = TRUE, mclust = TRUE,
  nbini = 20, matshape = FALSE)
}
\arguments{
\item{X}{the dataset (matrix)}

\item{nbclustmax}{max number of clusters in the gaussian mixtures}

\item{nbclustmin}{min number of clusters in the gaussian mixtures}

\item{verbose}{verbose or not}

\item{detailed}{boolean to give the details of the mixtures found}

\item{matshape}{boolean to give the detail in matricial shape}

\item{max}{boolean. Use an heuristic to shrink nbclustmax according to the number of individuals in the dataset}

\item{mclust}{boolean. Use mclust instead of Rmixmod}

\item{nbini}{number of initial points for Rmixmod}
}
\description{
Estimates the density of each covariates with gaussian mixture models and then gives the associated BIC.
}

