% Generated by roxygen2 (4.0.1): do not edit by hand
\name{Estep}
\alias{Estep}
\title{Imputation of missing values knowing alpha (E step of the EM)}
\usage{
Estep(X = X, alpha = alpha, M = NULL, Z = NULL, mixmod = mixmod,
  Zc = Zc, X1 = FALSE)
}
\arguments{
\item{X}{the dataset with missing values}

\item{M}{binary matrix X-sized (1=missing)}

\item{Ir}{vector of indices of endogenous covariates (redundent)}

\item{Ir}{the set of indices of the covariates on the left (computed once before EM to be efficient)}

\item{Zc}{profil (sum) colonne de Z}

\item{mixmod}{the matrix from mixmod}

\item{alpha}{matrix of the coefficients (p+1)xp format Matrix}

\item{X1}{boolean for alternative method}
}
\description{
Imputation of missing values knowing alpha (E step of the EM)
}

