% Generated by roxygen2 (4.0.1): do not edit by hand
\name{hatB}
\alias{hatB}
\title{Estimates B matrix}
\usage{
hatB(Z = Z, X = X, methode = 1)
}
\arguments{
\item{Z}{binary adjacency matrix of the structure (size p)}

\item{X}{the dataset}

\item{methode}{parameter for OLS (matrix inversion) methode_BIC  parameter for OLS (matrix inversion) 1:householderQr, 2:colPivHouseholderQr}
}
\description{
Estimates B matrix
}

