% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/matplot_zone.R
\name{matplot_zone}
\alias{matplot_zone}
\title{draws matplot with conditionnal background for easier comparison of curves.}
\usage{
matplot_zone(x = x, y = y, col = 1:6, alpha = 0.2, what = which.min,
  ylim = NULL, xlim = NULL, type = "p", xlab = NULL, ylab = NULL,
  main = NULL, ...)
}
\arguments{
\item{x}{the abscisses}

\item{y}{matrix of the curves (columns)}

\item{col}{list of colors (like in matplot)}

\item{alpha}{parameter for transparency of the background}

\item{what}{a function to choose a winner}

\item{ylim}{vector for vertical limits}

\item{xlim}{vector for horizontal limits}

\item{type}{the type of curve (like in matplot)}

\item{xlab}{(like in matplot)}

\item{ylab}{(like in matplot)}

\item{main}{the main title (like in matplot)}

\item{...}{Graphical parameters}
}
\description{
draws matplot with conditionnal background for easier comparison of curves.
}

