\name{newtheta}
\alias{newtheta}
\title{newtheta for Newton}
\usage{
newtheta(X = X, Z = Z, B = B, Sigma = Sigma, A = A, lambda = NULL,
  Atilde = Atilde, nbit = 1)
}
\arguments{
  \item{X}{dataset without intercept}

  \item{Z}{the structure (square without intercept)}

  \item{B}{the wheighted structure (p+1)xp}
}
\description{
newtheta for Newton
}

