\name{ProbaZ}
\alias{ProbaZ}
\title{Probability of Z without knowing the dataset. It also gives the exact number of binary nilpotent matrices of size p.}
\usage{
ProbaZ(Z = NULL, p = NULL, proba = FALSE, star = TRUE)
}
\arguments{
  \item{p}{the number of covariates}

  \item{Z}{the structure}

  \item{star}{gives the log pra under uniform law for p2}

  \item{proba}{gives the proba under the uniform law for Z}
}
\description{
Probability of Z without knowing the dataset. It also gives
the exact number of binary nilpotent matrices of size p.
}

