% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ProbaZ}
\alias{ProbaZ}
\title{Probability of Z without knowing the dataset. It also gives the exact number of binary nilpotent matrices of size p.}
\usage{
ProbaZ(Z = NULL, p = NULL, proba = FALSE, star = TRUE)
}
\arguments{
\item{Z}{binary adjacency matrix of the structure (size p)}

\item{p}{the number of covariates}

\item{proba}{gives the proba under the uniform law for Z. i FALSE and star=FALSE it gives the number of p-sized binary nilpotent matrices}

\item{star}{gives the log proba under uniform law for p2}
}
\description{
Probability of Z without knowing the dataset. It also gives the exact number of binary nilpotent matrices of size p.
}

