% Generated by roxygen2 (4.0.0): do not edit by hand
\name{readY}
\alias{readY}
\title{a summary-like function}
\usage{
readY(A = A, labels = NULL, X = NULL, intercept = TRUE, ANOVA = FALSE,
  print = FALSE, Y = NULL)
}
\arguments{
\item{A}{coefficient vector}

\item{labels}{name of the covariates}

\item{X}{the dataset (named) if labels is null}

\item{intercept}{boolean defining wether A contains an intercept or not}

\item{ANOVA}{boolean to add Anova test for each coefficient}

\item{print}{boolean to print ANOVA if computed}

\item{Y}{the response variable if ANOVA is computed}
}
\description{
a summary-like function
}

