% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/recursive_tree.R
\name{recursive_tree}
\alias{recursive_tree}
\title{decision tree in a recursive way}
\usage{
recursive_tree(data = data, Y = "Y", modele = NULL, kill = NULL,
  index = NULL, print = TRUE, plot = TRUE, main = NULL, sub = NULL,
  lang = c("en", "fr"))
}
\arguments{
\item{data}{the dataset including the response}

\item{Y}{the name of the response}

\item{modele}{vector of the names used for the last tree search (not necessarily choosen)}

\item{kill}{vector of the names to kill}

\item{index}{to give a number to the plot}

\item{print}{boolean to print the tree parameters}

\item{plot}{boolean to plot the tree}

\item{main}{the main title if plot=TRUE}

\item{sub}{the subtitle (if NULL it is automatically added)}

\item{lang}{the language for the automatic subtitle in the plot}
}
\description{
decision tree in a recursive way
}

