% Generated by roxygen2 (4.0.1): do not edit by hand
\name{rforge}
\alias{rforge}
\title{to upgrade a package to the lastest version on R-forge}
\usage{
rforge(packages = c("CorReg"), repos = "http://R-Forge.R-project.org")
}
\arguments{
\item{package}{vector of the name of the packages to upgrade}
}
\description{
to upgrade a package to the lastest version on R-forge
}

