% Generated by roxygen2 (4.0.1): do not edit by hand
\name{rforge}
\alias{rforge}
\title{Upgrades a package to the lastest version on R-forge}
\usage{
rforge(package = "CorReg")
}
\arguments{
\item{package}{name of the packages to upgrade}
}
\description{
Upgrades a package to the lastest version on R-forge
}

