% Generated by roxygen2 (4.0.1): do not edit by hand
\name{rforge}
\alias{rforge}
\title{to upgrade a package to the lastest version on R-forge}
\usage{
rforge(package = "CorReg")
}
\arguments{
\item{package}{the name of the package to upgrade}
}
\description{
to upgrade a package to the lastest version on R-forge
}

