% Generated by roxygen2 (4.0.0): do not edit by hand
\name{Terminator}
\alias{Terminator}
\title{Destructing values to have missing ones}
\usage{
Terminator(target = "Sarah Connor", wrath = 0.1)
}
\arguments{
\item{wrath}{the ratio of missing values in the output}

\item{target}{the dataset (matrix or data.frame) in which missing values will be made}
}
\description{
Destructing values to have missing ones
}

