% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/WhoIs.R
\name{WhoIs}
\alias{WhoIs}
\title{Give the partition implied by a structure}
\usage{
WhoIs(Z = Z, I3 = F, I2 = T, I1 = T)
}
\arguments{
\item{Z}{the structure (square matrix)}

\item{I1,I2,I3}{wanted outputs (I1= explicative covariates, I2=redundant covariates, I3=independent covariates)}
}
\description{
Give the partition implied by a structure
}

